+++ src/MyMediaLite/RatingPrediction/BiasedMatrixFactorization.cs
@@ -466,7 +466,7 @@ namespace MyMediaLite.RatingPrediction
 		}
 
 		///
-		public override double ComputeLoss()
+		public override float ComputeLoss()
 		{
 			double loss = 0;
 
@@ -497,7 +497,7 @@ namespace MyMediaLite.RatingPrediction
 				complexity += ratings.CountByItem[i] * BiasReg * Math.Pow(item_bias[i], 2);
 			}
 
-			return loss + complexity;
+			return (float) (loss + complexity);
 		}
 
 		///
